fixed some compile warnings for 64 bit architecture made compile on release mode

+++ TangoHapps/classes/project view/palette view/THPalette.h
@@ -84,7 +84,7 @@ You should have received a copy of the GNU General Public License along with thi
 -(void) removePaletteItem:(THPaletteItem*) paletteItem;
 -(void)removeAllPaletteItems;
 
--(int) getSize;
+-(NSUInteger) getSize;
 
 //temporary palette item
 -(void) temporaryAddPaletteItem:(THPaletteItem*) paletteItem;
